rename insert_into to insert
