+++ common/planners/src/plan_visitor.rs
@@ -120,7 +120,7 @@ pub trait PlanVisitor {
             PlanNode::Remote(plan) => self.visit_remote(plan),
             PlanNode::Having(plan) => self.visit_having(plan),
             PlanNode::Expression(plan) => self.visit_expression(plan),
-            PlanNode::InsertInto(plan) => self.visit_insert_into(plan),
+            PlanNode::Insert(plan) => self.visit_insert_into(plan),
             PlanNode::Copy(plan) => self.visit_copy(plan),
             PlanNode::ShowCreateTable(plan) => self.visit_show_create_table(plan),
             PlanNode::SubQueryExpression(plan) => self.visit_sub_queries_sets(plan),
