+++ query/src/interpreters/interpreter_insert.rs
@@ -22,9 +22,9 @@ use common_streams::DataBlockStream;
 use common_streams::SendableDataBlockStream;
 use futures::TryStreamExt;
 
-use super::interpreter_insert_into_with_stream::SendableWithSchema;
-use crate::interpreters::interpreter_insert_into_with_plan::InsertWithPlan;
-use crate::interpreters::interpreter_insert_into_with_stream::InsertWithStream;
+use super::interpreter_insert_with_stream::SendableWithSchema;
+use crate::interpreters::interpreter_insert_with_plan::InsertWithPlan;
+use crate::interpreters::interpreter_insert_with_stream::InsertWithStream;
 use crate::interpreters::Interpreter;
 use crate::interpreters::InterpreterPtr;
 use crate::sessions::QueryContext;
