+++ query/src/interpreters/interpreter_insert_with_stream.rs
