+++ query/src/servers/clickhouse/interactive_worker_base.rs
@@ -67,7 +67,7 @@ impl InteractiveWorkerBase {
         let plan = PlanParser::parse(query, ctx.clone()).await?;
 
         match plan {
-            PlanNode::InsertInto(insert) => Self::process_insert_query(insert, ch_ctx, ctx).await,
+            PlanNode::Insert(insert) => Self::process_insert_query(insert, ch_ctx, ctx).await,
             _ => {
                 let start = Instant::now();
                 let interpreter = InterpreterFactory::get(ctx.clone(), plan)?;
@@ -124,7 +124,7 @@ impl InteractiveWorkerBase {
             schema: sc,
         };
 
-        let interpreter = InterpreterFactory::get(ctx.clone(), PlanNode::InsertInto(insert))?;
+        let interpreter = InterpreterFactory::get(ctx.clone(), PlanNode::Insert(insert))?;
         let name = interpreter.name().to_string();
 
         let (mut tx, rx) = mpsc::channel(20);
